requirements.txt
flask==2.3.3
flask-cors==4.0.1
torch==2.0.1+cpu
torchvision==0.15.2+cpu
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4
gdown==5.2.0
gunicorn==22.0.0
transformers==4.44.2